requirements.txt
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
docutils==0.16
idna==2.10
importlib-metadata==3.7.2
keyring==23.0.0
numpy==1.21.5
opencv-python==4.5.4.60
packaging==20.9
Pillow==8.4.0
Pygments==2.8.1
pynput==1.7.5
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
pywin32==303
pywin32-ctypes==0.2.0
pyzbar==0.1.8
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
tqdm==4.59.0
twine==3.3.0
urllib3==1.26.3
webencodings==0.5.1
win10toast-click==0.1.2
wincertstore==0.2
zipp==3.4.1
